fix: move cleanup function and minor log fixes
